[issue#1]  Removing the function deviation from JPAExample class, and creating structured classes to attend this functionalities.
