--- src/main/java/br/com/bb/database/AccessDAO.java
@@ -1,12 +1,0 @@
-package br.com.bb.database;
-
-public interface AccessDAO {
-
-    Object save(Object object);
-
-    void update(Object object);
-
-    void delete(Object object);
-
-    void list(Object object);
-}
